reorganize sushi-hooks, update paths
